+++ src/main/java/com/spring/sharepod/controller/LikeRestController.java
@@ -3,7 +3,7 @@ package com.spring.sharepod.controller;
 
 import com.spring.sharepod.dto.request.LikeRequestDTO;
 import com.spring.sharepod.dto.response.BasicResponseDTO;
-import com.spring.sharepod.service.LikeService;
+import com.spring.sharepod.service.LikedService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -13,12 +13,12 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor //DI 형태, IoC 컨테이너 생성
 @RestController
 public class LikeRestController {
-    private final LikeService likeService;
+    private final LikedService likedService;
 
     //찜하기, 찜하기 취소
     @PostMapping("/like/{boardid}")
     public BasicResponseDTO likecontroll(@PathVariable Long boardid, @RequestBody LikeRequestDTO requestDTO){
-        return likeService.islikeservice(boardid,requestDTO);
+        return likedService.islikeservice(boardid,requestDTO);
     }
 
 }
