--- src/main/java/com/spring/sharepod/dto/request/LikedRequestDto.java
@@ -1,4 +1,0 @@
-package com.spring.sharepod.dto.request;
-
-public class LikedRequestDto {
-}
