+++ src/main/java/com/spring/sharepod/repository/LikedRepository.java
@@ -12,5 +12,5 @@ public interface LikedRepository extends JpaRepository<Liked, Long> {
     Liked findByUserAndBoard(User user, Board board);
 
     @Query("select c from Liked c where exists (select c from c where c.board.id=:boardid and c.user.id=:userid)")
-    Boolean findByLiked(Long boardid, Long userid);
+    Liked findByLiked(Long boardid, Long userid);
 }
